Add files via upload
I know it's a little late, but one of the papers I was using, well, something happened with the link and it sent me to the same website but a new page.... for $40. I had to find a new paper to use, as that's what I was basing most of this on. I will update my annotated bib shortly as well.